changed plot type

+++ RealTime/RT_b2s.py
@@ -442,11 +442,11 @@ def enhance_latest():
 
 
     fig,ax = plt.subplots(3,1,figsize=(20,10))
-    ax[0].imshow(cuts_beacon, cmap='gray', aspect='auto',interpolation='none',origin='upper', extent=[dates_beacon[0], dates_beacon[-1],elongations_beacon[0] , elongations_beacon[1]],vmin=vmin_beacon,vmax=vmax_beacon)
+    ax[0].imshow(cuts_beacon.astype(np.float32), cmap='gray', aspect='auto',interpolation='none',origin='upper', extent=[dates_beacon[0], dates_beacon[-1],elongations_beacon[0] , elongations_beacon[1]],vmin=vmin_beacon,vmax=vmax_beacon)
     ax[0].title.set_text('Beacon JPlot')
-    ax[1].imshow(cuts, cmap='gray', aspect='auto',interpolation='none',origin='upper', extent=[dates[0], dates[-1],elongations[0] , elongations[1]],vmin=vmin,vmax=vmax)
+    ax[1].imshow(cuts.astype(np.float32), cmap='gray', aspect='auto',interpolation='none',origin='upper', extent=[dates[0], dates[-1],elongations[0] , elongations[1]],vmin=vmin,vmax=vmax)
     ax[1].title.set_text('Enhanced Beacon JPlot')
-    ax[2].imshow(cuts_interpolated, cmap='gray', aspect='auto',interpolation='none',origin='upper', extent=[dates_interpolated[0], dates_interpolated[-1],elongations_interpolated[0] , elongations_interpolated[1]],vmin=vmin_interpolated,vmax=vmax_interpolated)
+    ax[2].imshow(cuts_interpolated.astype(np.float32), cmap='gray', aspect='auto',interpolation='none',origin='upper', extent=[dates_interpolated[0], dates_interpolated[-1],elongations_interpolated[0] , elongations_interpolated[1]],vmin=vmin_interpolated,vmax=vmax_interpolated)
     ax[2].title.set_text('Interpolated Enhanced Beacon JPlot')
     now  = datetime.now()
     # plt.savefig(str(now.year)+str('%02d' % now.month)+str('%02d' % now.day)+".png")
